deeper refactor of components
